requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
aiohttp==1.0.5            # via discord.py
async-timeout==1.1.0      # via aiohttp
beautifulsoup4==4.5.1
chardet==2.3.0            # via aiohttp
cleverbot==1.1.0
decorator==4.0.10         # via praw
discord.py==0.14.3
future==0.16.0            # via cleverbot
google==1.9.3
googlemaps==2.4.4
lxml==3.6.4
multidict==2.1.2          # via aiohttp
praw==3.5.0
requests==2.11.1
six==1.10                 # via praw
update-checker==0.11      # via praw
websockets==3.2           # via discord.py
